renaming my rna abstractor,, i think that abstract folder needs even more cleaning
